Move all table layout calculations from table_wraper.rs to table.rs
