+++ components/layout/table.rs
@@ -7,19 +7,20 @@
 #![deny(unsafe_code)]
 
 use app_units::Au;
-use block::{BlockFlow, CandidateBSizeIterator, ISizeAndMarginsComputer};
-use block::{ISizeConstraintInput, ISizeConstraintSolution};
+use block::{AbsoluteNonReplaced, BlockFlow, CandidateBSizeIterator, FloatNonReplaced};
+use block::{ISizeAndMarginsComputer, ISizeConstraintInput, ISizeConstraintSolution};
 use context::LayoutContext;
 use display_list::{BlockFlowDisplayListBuilding, BorderPaintingMode};
 use display_list::{DisplayListBuildState, StackingContextCollectionFlags, StackingContextCollectionState};
 use euclid::Point2D;
-use flow::{BaseFlow, EarlyAbsolutePositionInfo, Flow, FlowClass, ImmutableFlowUtils, GetBaseFlow, OpaqueFlow};
+use flow::{BaseFlow, EarlyAbsolutePositionInfo, Flow, FlowClass, FlowFlags};
+use flow::{ImmutableFlowUtils, GetBaseFlow, OpaqueFlow};
 use flow_list::{FlowListIterator, MutFlowListIterator};
 use fragment::{Fragment, FragmentBorderBoxIterator, Overflow};
 use gfx_traits::print_tree::PrintTree;
 use layout_debug;
 use model::{IntrinsicISizes, IntrinsicISizesContribution, MaybeAuto};
-use std::{cmp, fmt};
+use std::{cmp, fmt, ops::Add};
 use style::computed_values::{border_collapse, border_spacing, table_layout};
 use style::context::SharedStyleContext;
 use style::logical_geometry::LogicalSize;
@@ -87,10 +88,10 @@ impl TableFlow {
     /// a larger value than one of `self_inline_sizes`. Returns the minimum and preferred inline
     /// sizes.
     fn update_automatic_column_inline_sizes(
-            parent_inline_sizes: &mut Vec<ColumnIntrinsicInlineSize>,
-            child_cell_inline_sizes: &[CellIntrinsicInlineSize],
-            surrounding_size: Au)
-            -> IntrinsicISizes {
+        parent_inline_sizes: &mut Vec<ColumnIntrinsicInlineSize>,
+        child_cell_inline_sizes: &[CellIntrinsicInlineSize],
+        surrounding_size: Au
+    ) -> IntrinsicISizes {
         let mut total_inline_sizes = IntrinsicISizes {
             minimum_inline_size: surrounding_size,
             preferred_inline_size: surrounding_size,
@@ -158,12 +159,14 @@ impl TableFlow {
 
     /// Updates the minimum and preferred inline-size calculation for a single row. This is
     /// factored out into a separate function because we process children of rowgroups too.
-    fn update_column_inline_sizes_for_row(row: &TableRowFlow,
-                                          column_inline_sizes: &mut Vec<ColumnIntrinsicInlineSize>,
-                                          computation: &mut IntrinsicISizesContribution,
-                                          first_row: bool,
-                                          table_layout: TableLayout,
-                                          surrounding_inline_size: Au) {
+    fn update_column_inline_sizes_for_row(
+        row: &TableRowFlow,
+        column_inline_sizes: &mut Vec<ColumnIntrinsicInlineSize>,
+        computation: &mut IntrinsicISizesContribution,
+        first_row: bool,
+        table_layout: TableLayout,
+        surrounding_inline_size: Au
+    ) {
         // Read column inline-sizes from the table-row, and assign inline-size=0 for the columns
         // not defined in the column group.
         //
@@ -205,6 +208,13 @@ impl TableFlow {
         self.spacing().horizontal() * (num_columns as i32 + 1)
     }
 
+    fn compute_border_and_padding_of_table(&mut self) {
+        let available_inline_size = self.block_flow.base.block_container_inline_size;
+        self.block_flow.fragment.compute_border_and_padding(available_inline_size);
+        self.block_flow.fragment.compute_block_direction_margins(available_inline_size);
+        self.block_flow.fragment.compute_inline_direction_margins(available_inline_size);
+    }
+
     fn column_styles(&self) -> Vec<ColumnStyle> {
         let mut styles = vec![];
         for group in self.block_flow.base.child_iter()
@@ -234,6 +244,188 @@ impl TableFlow {
             }
         }
         styles
+    }
+
+    /// Calculates table column sizes for automatic layout per INTRINSIC § 4.3.
+    fn calculate_table_column_sizes_for_automatic_layout(
+        &mut self,
+        intermediate_column_inline_sizes: &mut [IntermediateColumnInlineSize]
+    ) {
+        let available_inline_size = self.available_inline_size();
+
+        // Compute all the guesses for the column sizes, and sum them.
+        let mut total_guess = AutoLayoutCandidateGuess::new();
+        let guesses: Vec<AutoLayoutCandidateGuess> =
+            self.column_intrinsic_inline_sizes.iter().map(|column_intrinsic_inline_size| {
+                let guess = AutoLayoutCandidateGuess::from_column_intrinsic_inline_size(
+                    column_intrinsic_inline_size,
+                    available_inline_size);
+                total_guess = &total_guess + &guess;
+                guess
+            }).collect();
+
+        // Assign inline sizes.
+        let selection = SelectedAutoLayoutCandidateGuess::select(&total_guess,
+                                                                 available_inline_size);
+        let mut total_used_inline_size = Au(0);
+        for (intermediate_column_inline_size, guess) in
+                intermediate_column_inline_sizes.iter_mut().zip(guesses.iter()) {
+            intermediate_column_inline_size.size = guess.calculate(selection);
+            intermediate_column_inline_size.percentage = 0.0;
+            total_used_inline_size = total_used_inline_size + intermediate_column_inline_size.size
+        }
+
+        // Distribute excess inline-size if necessary per INTRINSIC § 4.4.
+        //
+        // FIXME(pcwalton, spec): How do I deal with fractional excess?
+        let excess_inline_size = available_inline_size - total_used_inline_size;
+        if excess_inline_size > Au(0) && selection ==
+                SelectedAutoLayoutCandidateGuess::UsePreferredGuessAndDistributeExcessInlineSize {
+            let mut info = ExcessInlineSizeDistributionInfo::new();
+            for column_intrinsic_inline_size in &self.column_intrinsic_inline_sizes {
+                info.update(column_intrinsic_inline_size)
+            }
+
+            let mut total_distributed_excess_size = Au(0);
+            for (intermediate_column_inline_size, column_intrinsic_inline_size) in
+                    intermediate_column_inline_sizes.iter_mut()
+                                                    .zip(self.column_intrinsic_inline_sizes
+                                                             .iter()) {
+                info.distribute_excess_inline_size_to_column(intermediate_column_inline_size,
+                                                             column_intrinsic_inline_size,
+                                                             excess_inline_size,
+                                                             &mut total_distributed_excess_size)
+            }
+            total_used_inline_size = available_inline_size
+        }
+
+        self.set_inline_size(total_used_inline_size)
+    }
+
+    fn available_inline_size(&mut self) -> Au {
+        let available_inline_size = self.block_flow.fragment.border_box.size.inline;
+        let table_border_padding = self.block_flow.fragment.border_padding.inline_start_end();
+        let spacing = self.total_horizontal_spacing();
+
+        // FIXME(pcwalton, spec): INTRINSIC § 8 does not properly define how to compute this, but
+        // says "the basic idea is the same as the shrink-to-fit width that CSS2.1 defines". So we
+        // just use the shrink-to-fit inline size.
+        let available_inline_size = match self.block_flow.fragment.style().content_inline_size() {
+            LengthOrPercentageOrAuto::Auto => {
+                self.block_flow.get_shrink_to_fit_inline_size(available_inline_size) -
+                    table_border_padding
+            }
+            // FIXME(mttr): This fixes #4421 without breaking our current reftests, but I'm not
+            // completely sure this is "correct".
+            //
+            // That said, `available_inline_size` is, as far as I can tell, equal to the table's
+            // computed width property (W) and is used from this point forward in a way that seems
+            // to correspond with CSS 2.1 § 17.5.2.2 under "Column and caption widths influence the
+            // final table width as follows: …"
+            _ => available_inline_size,
+        };
+        available_inline_size - spacing
+    }
+
+    fn set_inline_size(&mut self, total_used_inline_size: Au) {
+        let table_border_padding = self.block_flow.fragment.border_padding.inline_start_end();
+        let spacing = self.total_horizontal_spacing();
+        self.block_flow.fragment.border_box.size.inline = total_used_inline_size +
+            table_border_padding + spacing;
+        self.block_flow.base.position.size.inline = total_used_inline_size +
+            table_border_padding + spacing + self.block_flow.fragment.margin.inline_start_end();
+
+        let writing_mode = self.block_flow.base.writing_mode;
+        let container_mode = self.block_flow.base.block_container_writing_mode;
+
+        if writing_mode.is_bidi_ltr() != container_mode.is_bidi_ltr() {
+            // If our "start" direction is different from our parent flow, then `border_box.start.i`
+            // depends on `border_box.size.inline`.
+            self.block_flow.fragment.border_box.start.i =
+                self.block_flow.base.block_container_inline_size -
+                self.block_flow.fragment.margin.inline_end -
+                self.block_flow.fragment.border_box.size.inline;
+        }
+    }
+
+    fn compute_used_inline_size(
+        &mut self,
+        shared_context: &SharedStyleContext,
+        parent_flow_inline_size: Au,
+        intermediate_column_inline_sizes: &[IntermediateColumnInlineSize]
+    ) {
+        let border_padding = self.block_flow.fragment.border_padding.inline_start_end();
+        let spacing = self.total_horizontal_spacing();
+        let minimum_width_of_all_columns =
+            intermediate_column_inline_sizes.iter()
+                                            .fold(border_padding + spacing,
+                                                  |accumulator, intermediate_column_inline_sizes| {
+                accumulator + intermediate_column_inline_sizes.size
+            });
+        let preferred_width_of_all_columns =
+            self.column_intrinsic_inline_sizes.iter()
+                                              .fold(border_padding + spacing,
+                                                    |accumulator, column_intrinsic_inline_sizes| {
+                accumulator + column_intrinsic_inline_sizes.preferred
+            });
+
+        // Delegate to the appropriate inline size computer to find the constraint inputs and write
+        // the constraint solutions in.
+        if self.block_flow.base.flags.is_float() {
+            let inline_size_computer = FloatedTable {
+                minimum_width_of_all_columns: minimum_width_of_all_columns,
+                preferred_width_of_all_columns: preferred_width_of_all_columns,
+                table_border_padding: border_padding,
+            };
+            let input =
+                inline_size_computer.compute_inline_size_constraint_inputs(&mut self.block_flow,
+                                                                           parent_flow_inline_size,
+                                                                           shared_context);
+
+            let solution = inline_size_computer.solve_inline_size_constraints(&mut self.block_flow,
+                                                                              &input);
+            inline_size_computer.set_inline_size_constraint_solutions(&mut self.block_flow,
+                                                                      solution);
+            inline_size_computer.set_inline_position_of_flow_if_necessary(&mut self.block_flow,
+                                                                          solution);
+            return
+        }
+
+        if !self.block_flow.base.flags.contains(FlowFlags::INLINE_POSITION_IS_STATIC) {
+            let inline_size_computer = AbsoluteTable {
+                minimum_width_of_all_columns: minimum_width_of_all_columns,
+                preferred_width_of_all_columns: preferred_width_of_all_columns,
+                table_border_padding: border_padding,
+            };
+            let input =
+                inline_size_computer.compute_inline_size_constraint_inputs(&mut self.block_flow,
+                                                                           parent_flow_inline_size,
+                                                                           shared_context);
+
+            let solution = inline_size_computer.solve_inline_size_constraints(&mut self.block_flow,
+                                                                              &input);
+            inline_size_computer.set_inline_size_constraint_solutions(&mut self.block_flow,
+                                                                      solution);
+            inline_size_computer.set_inline_position_of_flow_if_necessary(&mut self.block_flow,
+                                                                          solution);
+            return
+        }
+
+        let inline_size_computer = Table {
+            minimum_width_of_all_columns,
+            preferred_width_of_all_columns,
+            table_border_padding: border_padding,
+        };
+        let input =
+            inline_size_computer.compute_inline_size_constraint_inputs(&mut self.block_flow,
+                                                                       parent_flow_inline_size,
+                                                                       shared_context);
+
+        let solution = inline_size_computer.solve_inline_size_constraints(&mut self.block_flow,
+                                                                          &input);
+        inline_size_computer.set_inline_size_constraint_solutions(&mut self.block_flow, solution);
+        inline_size_computer.set_inline_position_of_flow_if_necessary(&mut self.block_flow,
+                                                                      solution);
     }
 }
 
@@ -323,12 +515,13 @@ impl Flow for TableFlow {
             let mut iterator = TableRowIterator::new(&mut self.block_flow.base).peekable();
             while let Some(row) = iterator.next() {
                 TableFlow::update_column_inline_sizes_for_row(
-                        row,
-                        &mut self.column_intrinsic_inline_sizes,
-                        &mut computation,
-                        first_row,
-                        self.table_layout,
-                        border_padding);
+                    row,
+                    &mut self.column_intrinsic_inline_sizes,
+                    &mut computation,
+                    first_row,
+                    self.table_layout,
+                    border_padding
+                );
                 if collapsing_borders {
                     let next_index_and_sibling = iterator.peek();
                     let next_collapsed_borders_in_block_direction =
@@ -382,8 +575,11 @@ impl Flow for TableFlow {
 
         let shared_context = layout_context.shared_context();
         // The position was set to the containing block by the flow's parent.
-        // FIXME: The code for distributing column widths should really be placed under table_wrapper.rs.
         let containing_block_inline_size = self.block_flow.base.block_container_inline_size;
+        if self.block_flow.base.flags.is_float() {
+            self.block_flow.float.as_mut().unwrap().containing_inline_size =
+                containing_block_inline_size;
+        }
 
         let mut constrained_column_inline_sizes_indices = vec![];
         let mut unspecified_inline_sizes_indices = vec![];
@@ -395,15 +591,25 @@ impl Flow for TableFlow {
             }
         }
 
-        let inline_size_computer = InternalTable;
-        inline_size_computer.compute_used_inline_size(&mut self.block_flow,
-                                                      shared_context,
-                                                      containing_block_inline_size);
+        let mut intermediate_column_inline_sizes: Vec<IntermediateColumnInlineSize> =
+            self.column_intrinsic_inline_sizes
+                .iter()
+                .map(|column_intrinsic_inline_size| {
+                    IntermediateColumnInlineSize {
+                        size: column_intrinsic_inline_size.minimum_length,
+                        percentage: column_intrinsic_inline_size.percentage,
+                    }
+                })
+                .collect();
 
-        let inline_start_content_edge = self.block_flow.fragment.border_padding.inline_start;
-        let inline_end_content_edge = self.block_flow.fragment.border_padding.inline_end;
+        self.compute_border_and_padding_of_table();
+        self.compute_used_inline_size(
+            shared_context,
+            containing_block_inline_size,
+            &intermediate_column_inline_sizes
+        );
+
         let padding_and_borders = self.block_flow.fragment.border_padding.inline_start_end();
-        let spacing_per_cell = self.spacing();
         let total_horizontal_spacing = self.total_horizontal_spacing();
         let content_inline_size = self.block_flow.fragment.border_box.size.inline -
             padding_and_borders - total_horizontal_spacing;
@@ -411,8 +617,6 @@ impl Flow for TableFlow {
 
         match self.table_layout {
             TableLayout::Fixed => {
-                self.column_computed_inline_sizes.clear();
-
                 // https://drafts.csswg.org/css2/tables.html#fixed-table-layout
                 for column_inline_size in &self.column_intrinsic_inline_sizes {
                     if column_inline_size.constrained {
@@ -454,12 +658,19 @@ impl Flow for TableFlow {
                     }
                 }
             }
-            _ => {
-                // The table wrapper already computed the inline-sizes and propagated them down
-                // to us.
+            TableLayout::Auto => {
+                self.calculate_table_column_sizes_for_automatic_layout(&mut intermediate_column_inline_sizes);
+                self.column_computed_inline_sizes = intermediate_column_inline_sizes.into_iter().map(|sizes| {
+                    ColumnComputedInlineSize {
+                        size: sizes.size,
+                    }
+                }).collect();
             }
         }
 
+        let inline_start_content_edge = self.block_flow.fragment.border_padding.inline_start;
+        let inline_end_content_edge = self.block_flow.fragment.border_padding.inline_end;
+        let spacing_per_cell = self.spacing();
         let column_computed_inline_sizes = &self.column_computed_inline_sizes;
         let collapsed_inline_direction_border_widths_for_table =
             &self.collapsed_inline_direction_border_widths_for_table;
@@ -580,39 +791,403 @@ impl fmt::Debug for TableFlow {
     }
 }
 
-/// Table, TableRowGroup, TableRow, TableCell types.
-/// Their inline-sizes are calculated in the same way and do not have margins.
-pub struct InternalTable;
+/// Returns the computed inline size of the table represented by `block`.
+///
+/// `table_border_padding` is the sum of the sizes of all border and padding in the inline
+/// direction of the table.
+fn initial_computed_inline_size(block: &mut BlockFlow,
+                                containing_block_inline_size: Au,
+                                minimum_width_of_all_columns: Au,
+                                preferred_width_of_all_columns: Au,
+                                table_border_padding: Au)
+                                -> MaybeAuto {
+    let inline_size_from_style = MaybeAuto::from_style(block.fragment.style.content_inline_size(),
+                                                       containing_block_inline_size);
+    match inline_size_from_style {
+        MaybeAuto::Auto => {
+            if preferred_width_of_all_columns + table_border_padding <= containing_block_inline_size {
+                MaybeAuto::Specified(preferred_width_of_all_columns + table_border_padding)
+            } else if minimum_width_of_all_columns > containing_block_inline_size {
+                MaybeAuto::Specified(minimum_width_of_all_columns)
+            } else {
+                MaybeAuto::Auto
+            }
+        }
+        MaybeAuto::Specified(inline_size_from_style) => {
+            MaybeAuto::Specified(cmp::max(
+                inline_size_from_style - table_border_padding,
+                minimum_width_of_all_columns
+            ))
+        }
+    }
+}
 
-impl ISizeAndMarginsComputer for InternalTable {
-    /// Compute the used value of inline-size, taking care of min-inline-size and max-inline-size.
-    ///
-    /// CSS Section 10.4: Minimum and Maximum inline-sizes
-    fn compute_used_inline_size(
-        &self,
-        block: &mut BlockFlow,
-        shared_context: &SharedStyleContext,
-        parent_flow_inline_size: Au
-    ) {
-        let mut input = self.compute_inline_size_constraint_inputs(block,
-                                                                   parent_flow_inline_size,
-                                                                   shared_context);
+struct Table {
+    minimum_width_of_all_columns: Au,
+    preferred_width_of_all_columns: Au,
+    table_border_padding: Au,
+}
 
-        // Tables are always at least as wide as their minimum inline size.
-        let minimum_inline_size =
-            block.base.intrinsic_inline_sizes.minimum_inline_size -
-            block.fragment.border_padding.inline_start_end();
-        input.available_inline_size = cmp::max(input.available_inline_size, minimum_inline_size);
-
-        let solution = self.solve_inline_size_constraints(block, &input);
-        self.set_inline_size_constraint_solutions(block, solution);
+impl ISizeAndMarginsComputer for Table {
+    fn initial_computed_inline_size(&self,
+                                    block: &mut BlockFlow,
+                                    parent_flow_inline_size: Au,
+                                    shared_context: &SharedStyleContext)
+                                    -> MaybeAuto {
+        let containing_block_inline_size =
+            self.containing_block_inline_size(block, parent_flow_inline_size, shared_context);
+        initial_computed_inline_size(block,
+                                     containing_block_inline_size,
+                                     self.minimum_width_of_all_columns,
+                                     self.preferred_width_of_all_columns,
+                                     self.table_border_padding)
     }
 
-    /// Solve the inline-size and margins constraints for this block flow.
-    fn solve_inline_size_constraints(&self, _: &mut BlockFlow, input: &ISizeConstraintInput)
+    fn solve_inline_size_constraints(&self,
+                                     block: &mut BlockFlow,
+                                     input: &ISizeConstraintInput)
                                      -> ISizeConstraintSolution {
-        ISizeConstraintSolution::new(input.available_inline_size, Au(0), Au(0))
+        self.solve_block_inline_size_constraints(block, input)
     }
+}
+
+struct FloatedTable {
+    minimum_width_of_all_columns: Au,
+    preferred_width_of_all_columns: Au,
+    table_border_padding: Au,
+}
+
+impl ISizeAndMarginsComputer for FloatedTable {
+    fn compute_border_and_padding(&self, block: &mut BlockFlow, containing_block_inline_size: Au) {
+        block.fragment.compute_border_and_padding(containing_block_inline_size)
+    }
+
+    fn initial_computed_inline_size(&self,
+                                    block: &mut BlockFlow,
+                                    parent_flow_inline_size: Au,
+                                    shared_context: &SharedStyleContext)
+                                    -> MaybeAuto {
+        let containing_block_inline_size =
+            self.containing_block_inline_size(block,
+                                              parent_flow_inline_size,
+                                              shared_context);
+        initial_computed_inline_size(block,
+                                     containing_block_inline_size,
+                                     self.minimum_width_of_all_columns,
+                                     self.preferred_width_of_all_columns,
+                                     self.table_border_padding)
+    }
+
+    fn solve_inline_size_constraints(&self,
+                                     block: &mut BlockFlow,
+                                     input: &ISizeConstraintInput)
+                                     -> ISizeConstraintSolution {
+        FloatNonReplaced.solve_inline_size_constraints(block, input)
+    }
+}
+
+struct AbsoluteTable {
+    minimum_width_of_all_columns: Au,
+    preferred_width_of_all_columns: Au,
+    table_border_padding: Au,
+}
+
+impl ISizeAndMarginsComputer for AbsoluteTable {
+    fn compute_border_and_padding(&self, block: &mut BlockFlow, containing_block_inline_size: Au) {
+        block.fragment.compute_border_and_padding(containing_block_inline_size)
+    }
+
+    fn initial_computed_inline_size(&self,
+                                    block: &mut BlockFlow,
+                                    parent_flow_inline_size: Au,
+                                    shared_context: &SharedStyleContext)
+                                    -> MaybeAuto {
+        let containing_block_inline_size =
+            self.containing_block_inline_size(block,
+                                              parent_flow_inline_size,
+                                              shared_context);
+        initial_computed_inline_size(block,
+                                     containing_block_inline_size,
+                                     self.minimum_width_of_all_columns,
+                                     self.preferred_width_of_all_columns,
+                                     self.table_border_padding)
+    }
+
+    fn containing_block_inline_size(&self,
+                                    block: &mut BlockFlow,
+                                    parent_flow_inline_size: Au,
+                                    shared_context: &SharedStyleContext)
+                                    -> Au {
+        AbsoluteNonReplaced.containing_block_inline_size(block,
+                                                         parent_flow_inline_size,
+                                                         shared_context)
+    }
+
+    fn solve_inline_size_constraints(&self,
+                                     block: &mut BlockFlow,
+                                     input: &ISizeConstraintInput)
+                                     -> ISizeConstraintSolution {
+        AbsoluteNonReplaced.solve_inline_size_constraints(block, input)
+    }
+
+    fn set_inline_position_of_flow_if_necessary(&self,
+                                                block: &mut BlockFlow,
+                                                solution: ISizeConstraintSolution) {
+        AbsoluteNonReplaced.set_inline_position_of_flow_if_necessary(block, solution);
+    }
+
+}
+
+/// The layout "guesses" defined in INTRINSIC § 4.3.
+struct AutoLayoutCandidateGuess {
+    /// The column inline-size assignment where each column is assigned its intrinsic minimum
+    /// inline-size.
+    minimum_guess: Au,
+
+    /// The column inline-size assignment where:
+    ///   * A column with an intrinsic percentage inline-size greater than 0% is assigned the
+    ///     larger of:
+    ///     - Its intrinsic percentage inline-size times the assignable inline-size;
+    ///     - Its intrinsic minimum inline-size;
+    ///   * Other columns receive their intrinsic minimum inline-size.
+    minimum_percentage_guess: Au,
+
+    /// The column inline-size assignment where:
+    ///   * Each column with an intrinsic percentage inline-size greater than 0% is assigned the
+    ///     larger of:
+    ///     - Its intrinsic percentage inline-size times the assignable inline-size;
+    ///     - Its intrinsic minimum inline-size;
+    ///   * Any other column that is constrained is assigned its intrinsic preferred inline-size;
+    ///   * Other columns are assigned their intrinsic minimum inline-size.
+    minimum_specified_guess: Au,
+
+    /// The column inline-size assignment where:
+    ///   * Each column with an intrinsic percentage inline-size greater than 0% is assigned the
+    ///     larger of:
+    ///     - Its intrinsic percentage inline-size times the assignable inline-size;
+    ///     - Its intrinsic minimum inline-size;
+    ///   * Other columns are assigned their intrinsic preferred inline-size.
+    preferred_guess: Au,
+}
+
+impl AutoLayoutCandidateGuess {
+    /// Creates a guess with all elements initialized to zero.
+    fn new() -> AutoLayoutCandidateGuess {
+        AutoLayoutCandidateGuess {
+            minimum_guess: Au(0),
+            minimum_percentage_guess: Au(0),
+            minimum_specified_guess: Au(0),
+            preferred_guess: Au(0),
+        }
+    }
+
+    /// Fills in the inline-size guesses for this column per INTRINSIC § 4.3.
+    fn from_column_intrinsic_inline_size(column_intrinsic_inline_size: &ColumnIntrinsicInlineSize,
+                                         assignable_inline_size: Au)
+                                         -> AutoLayoutCandidateGuess {
+        let minimum_percentage_guess = cmp::max(
+            assignable_inline_size.scale_by(column_intrinsic_inline_size.percentage),
+            column_intrinsic_inline_size.minimum_length
+        );
+        AutoLayoutCandidateGuess {
+            minimum_guess: column_intrinsic_inline_size.minimum_length,
+            minimum_percentage_guess: minimum_percentage_guess,
+            // FIXME(pcwalton): We need the notion of *constrainedness* per INTRINSIC § 4 to
+            // implement this one correctly.
+            minimum_specified_guess: if column_intrinsic_inline_size.percentage > 0.0 {
+                minimum_percentage_guess
+            } else if column_intrinsic_inline_size.constrained {
+                column_intrinsic_inline_size.preferred
+            } else {
+                column_intrinsic_inline_size.minimum_length
+            },
+            preferred_guess: if column_intrinsic_inline_size.percentage > 0.0 {
+                minimum_percentage_guess
+            } else {
+                column_intrinsic_inline_size.preferred
+            },
+        }
+    }
+
+    /// Calculates the inline-size, interpolating appropriately based on the value of `selection`.
+    ///
+    /// This does *not* distribute excess inline-size. That must be done later if necessary.
+    fn calculate(&self, selection: SelectedAutoLayoutCandidateGuess) -> Au {
+        match selection {
+            SelectedAutoLayoutCandidateGuess::UseMinimumGuess => self.minimum_guess,
+            SelectedAutoLayoutCandidateGuess::
+                    InterpolateBetweenMinimumGuessAndMinimumPercentageGuess(weight) => {
+                interp(self.minimum_guess, self.minimum_percentage_guess, weight)
+            }
+            SelectedAutoLayoutCandidateGuess::
+                    InterpolateBetweenMinimumPercentageGuessAndMinimumSpecifiedGuess(weight) => {
+                interp(self.minimum_percentage_guess, self.minimum_specified_guess, weight)
+            }
+            SelectedAutoLayoutCandidateGuess::
+                    InterpolateBetweenMinimumSpecifiedGuessAndPreferredGuess(weight) => {
+                interp(self.minimum_specified_guess, self.preferred_guess, weight)
+            }
+            SelectedAutoLayoutCandidateGuess::UsePreferredGuessAndDistributeExcessInlineSize => {
+                self.preferred_guess
+            }
+        }
+    }
+}
+
+impl<'a> Add for &'a AutoLayoutCandidateGuess {
+    type Output = AutoLayoutCandidateGuess;
+    #[inline]
+    fn add(self, other: &AutoLayoutCandidateGuess) -> AutoLayoutCandidateGuess {
+        AutoLayoutCandidateGuess {
+            minimum_guess: self.minimum_guess + other.minimum_guess,
+            minimum_percentage_guess:
+                self.minimum_percentage_guess + other.minimum_percentage_guess,
+            minimum_specified_guess: self.minimum_specified_guess + other.minimum_specified_guess,
+            preferred_guess: self.preferred_guess + other.preferred_guess,
+        }
+    }
+}
+
+/// The `CSSFloat` member specifies the weight of the smaller of the two guesses, on a scale from
+/// 0.0 to 1.0.
+#[derive(Clone, Copy, Debug, PartialEq)]
+enum SelectedAutoLayoutCandidateGuess {
+    UseMinimumGuess,
+    InterpolateBetweenMinimumGuessAndMinimumPercentageGuess(CSSFloat),
+    InterpolateBetweenMinimumPercentageGuessAndMinimumSpecifiedGuess(CSSFloat),
+    InterpolateBetweenMinimumSpecifiedGuessAndPreferredGuess(CSSFloat),
+    UsePreferredGuessAndDistributeExcessInlineSize,
+}
+
+impl SelectedAutoLayoutCandidateGuess {
+    /// See INTRINSIC § 4.3.
+    ///
+    /// FIXME(pcwalton, INTRINSIC spec): INTRINSIC doesn't specify whether these are exclusive or
+    /// inclusive ranges.
+    fn select(guess: &AutoLayoutCandidateGuess, assignable_inline_size: Au)
+              -> SelectedAutoLayoutCandidateGuess {
+        if assignable_inline_size < guess.minimum_guess {
+            SelectedAutoLayoutCandidateGuess::UseMinimumGuess
+        } else if assignable_inline_size < guess.minimum_percentage_guess {
+            let weight = weight(guess.minimum_guess,
+                                assignable_inline_size,
+                                guess.minimum_percentage_guess);
+            SelectedAutoLayoutCandidateGuess::InterpolateBetweenMinimumGuessAndMinimumPercentageGuess(weight)
+        } else if assignable_inline_size < guess.minimum_specified_guess {
+            let weight = weight(guess.minimum_percentage_guess,
+                                assignable_inline_size,
+                                guess.minimum_specified_guess);
+            SelectedAutoLayoutCandidateGuess::InterpolateBetweenMinimumPercentageGuessAndMinimumSpecifiedGuess(weight)
+        } else if assignable_inline_size < guess.preferred_guess {
+            let weight = weight(guess.minimum_specified_guess,
+                                assignable_inline_size,
+                                guess.preferred_guess);
+            SelectedAutoLayoutCandidateGuess::InterpolateBetweenMinimumSpecifiedGuessAndPreferredGuess(weight)
+        } else {
+            SelectedAutoLayoutCandidateGuess::UsePreferredGuessAndDistributeExcessInlineSize
+        }
+    }
+}
+
+/// Computes the weight needed to linearly interpolate `middle` between two guesses `low` and
+/// `high` as specified by INTRINSIC § 4.3.
+fn weight(low: Au, middle: Au, high: Au) -> CSSFloat {
+    (middle - low).to_f32_px() / (high - low).to_f32_px()
+}
+
+/// Linearly interpolates between two guesses, as specified by INTRINSIC § 4.3.
+fn interp(low: Au, high: Au, weight: CSSFloat) -> Au {
+    low + (high - low).scale_by(weight)
+}
+
+struct ExcessInlineSizeDistributionInfo {
+    preferred_inline_size_of_nonconstrained_columns_with_no_percentage: Au,
+    count_of_nonconstrained_columns_with_no_percentage: u32,
+    preferred_inline_size_of_constrained_columns_with_no_percentage: Au,
+    total_percentage: CSSFloat,
+    column_count: u32,
+}
+
+impl ExcessInlineSizeDistributionInfo {
+    fn new() -> ExcessInlineSizeDistributionInfo {
+        ExcessInlineSizeDistributionInfo {
+            preferred_inline_size_of_nonconstrained_columns_with_no_percentage: Au(0),
+            count_of_nonconstrained_columns_with_no_percentage: 0,
+            preferred_inline_size_of_constrained_columns_with_no_percentage: Au(0),
+            total_percentage: 0.0,
+            column_count: 0,
+        }
+    }
+
+    fn update(&mut self, column_intrinsic_inline_size: &ColumnIntrinsicInlineSize) {
+        if !column_intrinsic_inline_size.constrained &&
+                column_intrinsic_inline_size.percentage == 0.0 {
+            self.preferred_inline_size_of_nonconstrained_columns_with_no_percentage =
+                self.preferred_inline_size_of_nonconstrained_columns_with_no_percentage +
+                column_intrinsic_inline_size.preferred;
+            self.count_of_nonconstrained_columns_with_no_percentage += 1
+        }
+        if column_intrinsic_inline_size.constrained &&
+                column_intrinsic_inline_size.percentage == 0.0 {
+            self.preferred_inline_size_of_constrained_columns_with_no_percentage =
+                self.preferred_inline_size_of_constrained_columns_with_no_percentage +
+                column_intrinsic_inline_size.preferred
+        }
+        self.total_percentage += column_intrinsic_inline_size.percentage;
+        self.column_count += 1
+    }
+
+    /// Based on the information here, distributes excess inline-size to the given column per
+    /// INTRINSIC § 4.4.
+    ///
+    /// `#[inline]` so the compiler will hoist out the branch, which is loop-invariant.
+    #[inline]
+    fn distribute_excess_inline_size_to_column(
+        &self,
+        intermediate_column_inline_size: &mut IntermediateColumnInlineSize,
+        column_intrinsic_inline_size: &ColumnIntrinsicInlineSize,
+        excess_inline_size: Au,
+        total_distributed_excess_size: &mut Au
+    ) {
+        let proportion =
+            if self.preferred_inline_size_of_nonconstrained_columns_with_no_percentage > Au(0) {
+                // FIXME(spec, pcwalton): Gecko and WebKit do *something* here when there are
+                // nonconstrained columns with no percentage *and* no preferred width. What do they
+                // do?
+                if !column_intrinsic_inline_size.constrained &&
+                        column_intrinsic_inline_size.percentage == 0.0 {
+                    column_intrinsic_inline_size.preferred.to_f32_px() /
+                        self.preferred_inline_size_of_nonconstrained_columns_with_no_percentage
+                            .to_f32_px()
+                } else {
+                    0.0
+                }
+            } else if self.count_of_nonconstrained_columns_with_no_percentage > 0 {
+                1.0 / (self.count_of_nonconstrained_columns_with_no_percentage as CSSFloat)
+            } else if self.preferred_inline_size_of_constrained_columns_with_no_percentage >
+                    Au(0) {
+                column_intrinsic_inline_size.preferred.to_f32_px() /
+                    self.preferred_inline_size_of_constrained_columns_with_no_percentage.to_f32_px()
+            } else if self.total_percentage > 0.0 {
+                column_intrinsic_inline_size.percentage / self.total_percentage
+            } else {
+                1.0 / (self.column_count as CSSFloat)
+            };
+
+        // The `min` here has the effect of throwing away fractional excess at the end of the
+        // table.
+        let amount_to_distribute = cmp::min(excess_inline_size.scale_by(proportion),
+                                       excess_inline_size - *total_distributed_excess_size);
+        *total_distributed_excess_size = *total_distributed_excess_size + amount_to_distribute;
+        intermediate_column_inline_size.size = intermediate_column_inline_size.size +
+            amount_to_distribute
+    }
+}
+
+/// An intermediate column size assignment.
+struct IntermediateColumnInlineSize {
+    size: Au,
+    percentage: f32,
 }
 
 /// Information about the intrinsic inline sizes of columns within a table.
